+++ src/app/features/templating/templateValuesSrv.js
@@ -82,7 +82,7 @@ function (angular, _, kbn) {
     this._updateNonQueryVariable = function(variable) {
       // extract options in comma seperated string
       variable.options = _.map(variable.query.split(/[,]+/), function(text) {
-        return { text: text, value: text };
+        return { text: text.trim(), value: text.trim() };
       });
 
       if (variable.type === 'interval') {