+++ src/test/specs/graph-tooltip-specs.js
@@ -19,6 +19,8 @@ define([
       tooltip:  {
         shared: true
       },
+      legend: {
+      },
       stack: false
     };